Move view instance variable logic to helpers
We had helper methods that returned the `view` instance variable, but in
the `view_mode` partials we were using this instance variable directly.
As a previous step to remove the code duplication, we're moving this
logic from the controllers; sharing code in the controllers would
require a concern and in this case it isn't worth it.